Render into a multisampled framebuffer
Because we want to implement a better algorithm to deal with sorting of
transparent fragments, we start by rendering the scene into an offscreen
buffer that has multisampling enabled.

We allocate an oversized framebuffer/texture that is the size of the
monitor. Then to deal with changes in the window size we render into a
smaller rectangle of the the full framebuffer's texture. Then we render
a quad that is transformed to match the original positions of
everything. We do this because rendering to the full oversized
framebuffer and then scaling back into the window causes some artifacts
that make it look like we are not using multisampling (I left an #ifdef
to show the artifacts).

I haven't checked if this transform produces pixel perfect results and
it may be the case that some firtering is triggered. Maybe a better
method is needed to blend pixel perfect textures coming for example
from cairo.

This method avoids resizing the frambuffer's texture, OpenGL is supposed
to allow this but at the moment I don't know how both alternatives
compare performance wise. Also, I chose this because there were coments
online about texture resizing not being well supported in some OpenGL
implementations. I don't have enough information to make a concrete
decission so we do this for now.